Add a form to submit the appointment date.
In order to submit the appointment date, a Save Changes button was needed to submit the selected date along with the property id.

The date picker will also not allow you to pick a date before an hour after the page was loaded. This makes setting an appointment more realistic.