Create a base class for the command handlers.
Each of the command handlers had things in common; they each had a DbContext and they each had a void Handle method.

In order to make this more of a pattern, an abstract base class was created with a constructor for the DbContext and an abstract void Handle method.

The base class is generic to allow the command to be passed in.