Add functionality to accept and reject an appointment.
Similar to making an offer, a seller should be able to accept and reject an appointment made by a buyer.

Added a new link for both the buyer and seller called "My Appointments" which will show the appointments linked to each user.

Added logic to redirect to an invalid appointment if a buyer tried to schedule an appointment within an hour of an already accepted appointment.